requirements.txt
APScheduler==3.10.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Login==0.6.3
gunicorn==21.2.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pybit==5.6.2
pycryptodome==3.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
